change delete route to admin/delete

+++ routes/web.php
@@ -21,4 +21,4 @@ Route::get('/', [EventController::class, 'index'])->name('home');
 Auth::routes();
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
-Route::delete('/delete/{id}', [EventController::class, 'destroy'])->name('delete');
+Route::delete('admin/delete/{id}', [EventController::class, 'destroy'])->name('delete');
